Rename distribute_tests.py to distribute.py
